+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/delegation/DelegateGet.scala
@@ -50,7 +50,7 @@ case class DelegateIds(value: List[UnparsedDelegateId])
 case class DelegateGetRequest(accountId: AccountId,
                               ids: Option[DelegateIds],
                               properties: Option[Properties]) extends WithAccountId with GetRequest {
-  override def idCount: Option[Long] = ids.map(_.value).map(_.size)
+  override def idCount: Option[Int] = ids.map(_.value).map(_.size)
 }
 case class Delegate(id: DelegationId,
                     username: Username) {
