+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/delegation/DelegatedAccountGet.scala
@@ -34,7 +34,7 @@ object DelegatedAccountGet {
 case class DelegatedAccountGetRequest(accountId: AccountId,
                                       ids: Option[DelegateIds],
                                       properties: Option[Properties]) extends WithAccountId with GetRequest {
-  override def idCount: Option[Long] = ids.map(_.value).map(_.size)
+  override def idCount: Option[Int] = ids.map(_.value).map(_.size)
 }
 
 case class DelegatedAccountNotFound(value: Set[UnparsedDelegateId])
