+++ server/protocols/jmap-rfc-8621/src/main/scala/org/apache/james/jmap/delegation/DelegatedAccountSet.scala
@@ -5,7 +5,7 @@ import org.apache.james.jmap.method.{SetRequest, WithAccountId}
 
 case class DelegatedAccountSetRequest(accountId: AccountId,
                                       destroy: Option[Seq[UnparsedDelegateId]]) extends WithAccountId with SetRequest {
-  override def idCount: Long = destroy.map(_.size).getOrElse(0)
+  override def idCount: Int = destroy.map(_.size).getOrElse(0)
 }
 
 case class DelegatedAccountSetResponse(accountId: AccountId,
